Fix: CFs were inserted into items StereologicalInclusion extension when it shouldn't preventing file saving later.
- Fix: check item's stack and compare it with CF's stack to avoid a wrong inclusion.
- Fix: clear old infocache information from StereologicalInclusion extension from being reported when saving a file.

+++ App/ToolGroups/Edit/CODERefineWidget.cpp
@@ -142,7 +142,7 @@ CODERefineWidget::~CODERefineWidget()
 void CODERefineWidget::onRadiusModified(int value)
 {
   m_gui->radius->setValue(value);
-  m_gui->apply->setEnabled(value == m_filter->radius());
+  m_gui->apply->setEnabled(value == static_cast<int>(m_filter->radius()));
 }
 
 //----------------------------------------------------------------------------
